banded to sparsetools; import with *
